requirements.txt
python-dotenv
langchain
langchain-openai
langchain-chroma
tiktoken
black
streamlit